requirements.txt
backcall==0.1.0
beautifulsoup4==4.9.1
bootstrap4==0.1.0
colorama==0.4.3
decorator==4.4.2
Django==2.1.15
django-bootstrap-pagination==1.7.1
django-bootstrap4==1.1.1
django-extensions==2.2.9
djangorestframework==3.11.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
Pygments==2.6.1
pytz==2020.1
six==1.14.0
soupsieve==2.0.1
traitlets==4.3.3
wcwidth==0.1.9
